refactor: 파일 구조 변경
Endpoint 파일을 service 디렉토리로 변경함.
